+++ tests/test_selectors.py
@@ -181,6 +181,8 @@ class TestTextQuoteSelectors:
 
 
 class TestCreateTextPositionSelectors:
+    great_text = "Here is some great text to excerpt."
+
     def test_dump_position_selector(self):
         selector = TextPositionSelector(start=5, end=12)
         dumped = selector.dump()
@@ -252,28 +254,38 @@ class TestCreateTextPositionSelectors:
             TextPositionSelector("[0,0)")
 
     def test_create_selector_from_phrase_and_number(self):
-        text = "Here is some great text to excerpt."
-        result = TextPositionSelector.from_text(text=text, start="some", end=23)
+        result = TextPositionSelector.from_text(
+            text=self.great_text, start="some", end=23
+        )
         assert result.start == 8
         assert result.end == 23
-        assert result.passage(text) == "some great text"
+        assert result.passage(self.great_text) == "some great text"
 
     def test_create_selector_from_phrases(self):
-        text = "Here is some great text to excerpt."
-        result = TextPositionSelector.from_text(text=text, start="some", end="text")
+        result = TextPositionSelector.from_text(
+            text=self.great_text, start="some", end="text"
+        )
         assert result.start == 8
         assert result.end == 23
-        assert result.passage(text) == "some great text"
+        assert result.passage(self.great_text) == "some great text"
 
     def test_wrong_start_phrase_selector(self):
-        text = "Here is some great text to excerpt."
         with pytest.raises(TextSelectionError):
-            TextPositionSelector.from_text(text=text, start="snipe", end=23)
+            TextPositionSelector.from_text(text=self.great_text, start="snipe", end=23)
 
     def test_wrong_end_phrase_selector(self):
-        text = "Here is some great text to excerpt."
         with pytest.raises(TextSelectionError):
-            TextPositionSelector.from_text(text=text, start=8, end="wild goose")
+            TextPositionSelector.from_text(
+                text=self.great_text, start=8, end="wild goose"
+            )
+
+    def test_phrase_selector_start_only(self):
+        selector = TextPositionSelector.from_text(text=self.great_text, start=8)
+        assert selector.passage(self.great_text) == "some great text to excerpt."
+
+    def test_phrase_selector_end_only(self):
+        selector = TextPositionSelector.from_text(text=self.great_text, end=23)
+        assert selector.passage(self.great_text) == "Here is some great text"
 
 
 class TestCombineTextPositionSelectors: