+++ changelog.md
@@ -1,3 +1,7 @@
+dev
+------------------
+- add TextPositionSelector.from_text constructor
+
 0.4.4 (2020-01-25)
 ------------------
 - provide "missing" instead of "optional" argument for marshmallow schema